web/requirements/common.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.10
django-configurations==1.0.0
django-dotenv==1.4.1
django-filter==0.15.3
djangorestframework==3.5.3
inflection==0.3.0
Markdown==2.6.7
psycopg2==2.6.2
pyyaml==3.11
whitenoise==3.2.1
